+++ src/components/PropProvider/docs/propConnect.md
@@ -1,6 +1,6 @@
 # propConnect
 
-Higher-order function that "connects" a React component with Blue's config, generated by the [PropProvider](./Provider.md). This pattern is similar to [Redux's connect](https://github.com/reduxjs/react-redux).
+Higher-order function that "connects" a React component with HSDS React's config, generated by the [PropProvider](./Provider.md). This pattern is similar to [Redux's connect](https://github.com/reduxjs/react-redux).
 
 ## Examples
