+++ src/components/PropProvider/index.ts
@@ -1,6 +1,8 @@
 import PropProvider from './PropProvider'
 export { default as Provider } from './Provider'
 export { default as propConnect } from './propConnect'
+export { default as contextConnect } from './contextConnect'
+
 export {
   APPS,
   getGlobalApp,
