+++ src/components/PropProvider/utils.ts
@@ -176,3 +176,13 @@ export function getConfigPropsFromArray(
     return remappedProps
   }, {})
 }
+
+/**
+ * Checks to see of a component is a Stateless Functional Component.
+ *
+ * @param {any} Component The component to check.
+ * @returns {boolean} The result.
+ */
+export function isStateless(Component: any): boolean {
+  return !isString(Component) && !Component.prototype.render
+}
